More refactor to pic and interrupts
